Convert j1client to typescript
