+++ package.json
@@ -28,19 +28,18 @@
     "apollo-cache-inmemory": "^1.5.1",
     "apollo-client": "^2.5.1",
     "apollo-link": "^1.2.11",
-    "apollo-link-batch-http": "^1.2.11",
-    "apollo-link-http": "^1.5.12",
+    "apollo-link-batch-http": "^1.2.13",
     "apollo-link-retry": "^2.2.13",
     "commander": "^2.14.1",
     "file-set": "^2.0.0",
     "fs-extra": "^5.0.0",
-    "graphql": "^14.1.1",
+    "graphql": "^14.6.0",
     "graphql-tag": "^2.10.1",
     "inquirer": "^5.1.0",
     "js-yaml": "^3.13.1",
     "moment": "^2.22.2",
     "mustache": "^2.3.0",
-    "node-fetch": "^2.3.0",
+    "node-fetch": "^2.6.0",
     "p-all": "^2.1.0",
     "p-map": "^1.2.0",
     "papaparse": "^4.6.3",
@@ -52,6 +51,8 @@
     "valid-url": "^1.0.9"
   },
   "devDependencies": {
+    "@types/jest": "^25.1.4",
+    "@types/node": "^13.9.8",
     "dotenv": "^7.0.0",
     "husky": "^2.4.0",
     "jest": "^24.9.0",
